feat: increase max item stack size from 100 to 1000
- Change item count storage from uint8_t to uint16_t in Item class
- Update all hardcoded stack limit checks (100 → 1000) across
  container, game, player, tile, luascript, npc, and iomarket
- Update network protocol to send item count as uint16_t instead
  of a single byte (addByte → add<uint16_t>) in NetworkMessage
- Update server protocol parsers (parseThrow, parseLookInShop,
  parseEquipObject) to read uint16_t counts from client packets
- Enable GameCountU16 feature in client features.lua so the client
  sends and receives counts as 2-byte values

+++ src/item.h
@@ -1103,7 +1103,7 @@ class Item : virtual public Thing, public SharedObject
 			return count;
 		}
 	
-		void setItemCount(uint8_t n) {
+		void setItemCount(uint16_t n) {
 			count = n;
 		}
 
@@ -1347,7 +1347,7 @@ class Item : virtual public Thing, public SharedObject
 
 	private:
 		uint16_t imbuementSlots = 0;
-		uint8_t count = 1; // number of stacked items
+		uint16_t count = 1; // number of stacked items
 		bool loadedFromMap = false;
         std::string getWeightDescription(uint32_t weight) const;
 };
